[Refactor] 예외 처리 패키지 추가 및 파일 이동 - 같은 유형의 예외끼리 묶기 위해서 예외 처리 클래스 파일들의 패키지 추가 및 이동
