+++ backend/ricecake/src/main/java/project/ricecake/error/exception/duplicate/DuplicateException.java
@@ -1,6 +1,7 @@
-package project.ricecake.error.exception;
+package project.ricecake.error.exception.duplicate;
 
 import project.ricecake.error.ErrorCode;
+import project.ricecake.error.exception.BusinessException;
 
 public class DuplicateException extends BusinessException {
     public DuplicateException(ErrorCode errorCode) {
