+++ backend/ricecake/src/main/java/project/ricecake/error/exception/notfound/BoardNotFoundException.java
@@ -1,4 +1,4 @@
-package project.ricecake.error.exception;
+package project.ricecake.error.exception.notfound;
 
 import project.ricecake.error.ErrorCode;
