+++ backend/ricecake/src/main/java/project/ricecake/error/exception/notfound/NotFoundException.java
@@ -1,6 +1,7 @@
-package project.ricecake.error.exception;
+package project.ricecake.error.exception.notfound;
 
 import project.ricecake.error.ErrorCode;
+import project.ricecake.error.exception.BusinessException;
 
 public class NotFoundException extends BusinessException {
     public NotFoundException(ErrorCode errorCode) {
