+++ backend/ricecake/src/main/java/project/ricecake/error/exception/unauthorized/UnauthorizedException.java
@@ -1,6 +1,7 @@
-package project.ricecake.error.exception;
+package project.ricecake.error.exception.unauthorized;
 
 import project.ricecake.error.ErrorCode;
+import project.ricecake.error.exception.BusinessException;
 
 public class UnauthorizedException extends BusinessException {
     public UnauthorizedException(ErrorCode errorCode) {
